requirements.txt
paho-mqtt==1.6.1
sqlalchemy==2.0.23
polars==0.20.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
uuid==1.30
